+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+1.3.2 (2021-08-30)
+------------------
+
+- starlette integration added
+- django integration added
+- sub endpoints support implemented
+
+
 1.3.1 (2021-08-24)
 ------------------